feat: 보내요 봉투 세부 내역 서버 연동

+++ data/src/main/java/com/susu/data/remote/api/EnvelopesService.kt
@@ -32,7 +32,7 @@ interface EnvelopesService {
 
     @GET("envelopes")
     suspend fun getEnvelopesHistoryList(
-        @Query("friendIds") friendIds: List<Int>?,
+        @Query("friendIds") friendIds: List<Long>?,
         @Query("ledgerId") ledgerId: Int?,
         @Query("types") types: List<String>?,
         @Query("include") include: List<String>?,
